chore: add Provider component as global decorator to DF Storybook (#889)
